+++ packages/turf-helpers/AllGeoJSON.ts
@@ -1,0 +1,12 @@
+import type {
+  Feature,
+  FeatureCollection,
+  Geometry,
+  GeometryCollection,
+} from "geojson";
+
+export type AllGeoJSON =
+  | Feature
+  | FeatureCollection
+  | Geometry
+  | GeometryCollection;
