+++ packages/turf-helpers/AreaUnits.ts
@@ -1,0 +1,6 @@
+import { Units } from "./Units.js";
+
+export type AreaUnits =
+  | Exclude<Units, "radians" | "degrees">
+  | "acres"
+  | "hectares";
