+++ packages/turf-helpers/Coord.ts
@@ -1,0 +1,3 @@
+import type { Feature, Point, Position } from "geojson";
+
+export type Coord = Feature<Point> | Point | Position;
