+++ packages/turf-helpers/Corners.ts
@@ -1,0 +1,1 @@
+export type Corners = "sw" | "se" | "nw" | "ne" | "center" | "centroid";
