+++ packages/turf-helpers/Grid.ts
@@ -1,0 +1,1 @@
+export type Grid = "point" | "square" | "hex" | "triangle";
