+++ packages/turf-helpers/Lines.ts
@@ -1,0 +1,8 @@
+import type {
+  LineString,
+  MultiLineString,
+  MultiPolygon,
+  Polygon,
+} from "geojson";
+
+export type Lines = LineString | MultiLineString | Polygon | MultiPolygon;
