+++ packages/turf-helpers/Units.ts
@@ -1,0 +1,16 @@
+export type Units =
+  | "meters"
+  | "metres"
+  | "millimeters"
+  | "millimetres"
+  | "centimeters"
+  | "centimetres"
+  | "kilometers"
+  | "kilometres"
+  | "miles"
+  | "nauticalmiles"
+  | "inches"
+  | "yards"
+  | "feet"
+  | "radians"
+  | "degrees";
