+++ packages/turf-helpers/earthRadius.ts
@@ -1,0 +1,7 @@
+/**
+ * Earth Radius used with the Harvesine formula and approximates using a spherical (non-ellipsoid) Earth.
+ *
+ * @memberof helpers
+ * @type {number}
+ */
+export const earthRadius = 6371008.8;
