+++ packages/turf-helpers/geometry.ts
@@ -1,0 +1,50 @@
+import { lineString } from "./lineString.js";
+import { multiLineString } from "./multiLineString.js";
+import { multiPoint } from "./multiPoint.js";
+import { multiPolygon } from "./multiPolygon.js";
+import { point } from "./point.js";
+import { polygon } from "./polygon.js";
+
+/**
+ * Creates a GeoJSON {@link Geometry} from a Geometry string type & coordinates.
+ * For GeometryCollection type use `helpers.geometryCollection`
+ *
+ * @name geometry
+ * @param {string} type Geometry Type
+ * @param {Array<any>} coordinates Coordinates
+ * @param {Object} [options={}] Optional Parameters
+ * @returns {Geometry} a GeoJSON Geometry
+ * @example
+ * var type = "Point";
+ * var coordinates = [110, 50];
+ * var geometry = turf.geometry(type, coordinates);
+ * // => geometry
+ */
+export function geometry(
+  type:
+    | "Point"
+    | "LineString"
+    | "Polygon"
+    | "MultiPoint"
+    | "MultiLineString"
+    | "MultiPolygon",
+  coordinates: any[],
+  _options: Record<string, never> = {}
+) {
+  switch (type) {
+    case "Point":
+      return point(coordinates).geometry;
+    case "LineString":
+      return lineString(coordinates).geometry;
+    case "Polygon":
+      return polygon(coordinates).geometry;
+    case "MultiPoint":
+      return multiPoint(coordinates).geometry;
+    case "MultiLineString":
+      return multiLineString(coordinates).geometry;
+    case "MultiPolygon":
+      return multiPolygon(coordinates).geometry;
+    default:
+      throw new Error(type + " is invalid");
+  }
+}
