+++ packages/turf-helpers/lengthToRadians.ts
@@ -1,0 +1,23 @@
+import type { Units } from "./Units.js";
+import { factors } from "./factors.js";
+
+/**
+ * Convert a distance measurement (assuming a spherical Earth) from a real-world unit into radians
+ * Valid units: miles, nauticalmiles, inches, yards, meters, metres, kilometers, centimeters, feet
+ *
+ * @name lengthToRadians
+ * @param {number} distance in real units
+ * @param {string} [units="kilometers"] can be degrees, radians, miles, inches, yards, metres,
+ * meters, kilometres, kilometers.
+ * @returns {number} radians
+ */
+export function lengthToRadians(
+  distance: number,
+  units: Units = "kilometers"
+): number {
+  const factor = factors[units];
+  if (!factor) {
+    throw new Error(units + " units is invalid");
+  }
+  return distance / factor;
+}
