+++ packages/turf-helpers/radiansToDegrees.ts
@@ -1,0 +1,11 @@
+/**
+ * Converts an angle in radians to degrees
+ *
+ * @name radiansToDegrees
+ * @param {number} radians angle in radians
+ * @returns {number} degrees between 0 and 360 degrees
+ */
+export function radiansToDegrees(radians: number): number {
+  const degrees = radians % (2 * Math.PI);
+  return (degrees * 180) / Math.PI;
+}
