+++ packages/turf-helpers/round.ts
@@ -1,0 +1,20 @@
+/**
+ * Round number to precision
+ *
+ * @param {number} num Number
+ * @param {number} [precision=0] Precision
+ * @returns {number} rounded number
+ * @example
+ * turf.round(120.4321)
+ * //=120
+ *
+ * turf.round(120.4321, 2)
+ * //=120.43
+ */
+export function round(num: number, precision = 0): number {
+  if (precision && !(precision >= 0)) {
+    throw new Error("precision must be a positive number");
+  }
+  const multiplier = Math.pow(10, precision || 0);
+  return Math.round(num * multiplier) / multiplier;
+}
